% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Funct.R
\name{calculate_filter_function}
\alias{calculate_filter_function}
\title{Filtering function}
\usage{
calculate_filter_function(genes_case_full_data, survival_time, survival_event)
}
\arguments{
\item{genes_case_full_data}{Input matrix whose columns correspond to the patients
and rows to the genes, having selected only the columns belonging to disease
samples and having selected the rows corresponding to the genes selected
by the function \code{gene_selection_fun_filt}. The names of the rows must
be the names of the genes.}

\item{survival_time}{Numeric vector that includes time to the event
information after removal of NAs corresponding to healthy tissue samples
(case_tag == control_tag).}

\item{survival_event}{Numeric vector that indicates if relapse or death
have been produced (0 and 1s) after removal of NAs corresponding to healthy
tissue samples (case_tag == control_tag).}
}
\value{
A numeric vector including the values produced by the function
for each sample in the dataset \code{genes_case_full_data}.
}
\description{
A filtering function for Mapper that projects $$R$^n$ into $R$.
It calculates for each column of the matrix (each patient), its value
of the filtering function.
Specifically, it computes a multivariate Cox proportional hazard model using
all previously selected genes as predictor variables. Subsequently, it
calculates for each patient using this model their linear predictor,
which corresponds to the exponent of their proportionality constant.
}
