% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_case_tag}
\alias{check_case_tag}
\title{check_case_tag}
\usage{
check_case_tag(full_data, case_tag, control_tag)
}
\arguments{
\item{full_data}{The genes of the full_data}

\item{case_tag}{The tag of the healthy sample (healthy or not).}

\item{control_tag}{Tag of the healthy sample.E.g. "T"}
}
\value{
A list containing the \code{control_tag} (the tag of the healthy
sample), and the full_data and case_tag updated (samples with NA's in the
case_tag vector deleted).
}
\description{
Checking \code{case_tag} introduced in the \code{dsga} object.
Function used internally in the funtion \code{dsga}.
If there are NAs in the case_tag, the corresponding samples are removed in
\code{full_data}.
}
