% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_plot_vectors}
\alias{check_plot_vectors}
\title{check_plot_vectors}
\usage{
check_plot_vectors(mapper_object, vector, class_vector = "character")
}
\arguments{
\item{mapper_object}{A \code{mapper_obj} object.}

\item{vector}{Vector containing the variable to be used for representation.
It must contain a single value per sample.}

\item{class_vector}{"numeric" or "character". Variable indicating the type
of vector depending on if it is to be used in \code{plot_mapper} or in
\code{plot_mapper_with_pie_chart}. "character" by default.}
}
\value{
No return, only checking function
}
\description{
Check the vectors for the functions plot_mapper and plot_mapper_with_pie_chart.
}
