% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_surv_vectors}
\alias{check_surv_vectors}
\title{check_surv_vectors}
\usage{
check_surv_vectors(full_data, survival_time, survival_event)
}
\arguments{
\item{full_data}{The genes of the full_data}

\item{survival_time}{Time between disease diagnosis and event (if there was
no event until the end of follow-up).}

\item{survival_event}{\code{logical}. Whether or not the event has occurred.}
}
\value{
No return, only checking function
}
\description{
Checking the \code{survival_time} and \code{survival_event}
vectors introduced in the \code{gene_selection object}.
Function used internally in the funtion \code{gene_selection}, if the
object is a \code{dsga_object}.
}
