% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_vectors}
\alias{check_vectors}
\title{check_vectors}
\usage{
check_vectors(full_data, survival_time, survival_event, case_tag, control_tag)
}
\arguments{
\item{full_data}{The genes of the full_data}

\item{survival_time}{Time between disease diagnosis and event (if there was
no event until the end of follow-up).}

\item{survival_event}{\code{logical}. Whether or not the event has occurred.}

\item{case_tag}{The tag of the healthy sample (healthy or not).}

\item{control_tag}{Tag of the healthy sample.E.g. "T"}
}
\value{
A list containing the \code{control_tag} (the tag of the healthy
sample), and the \code{full_data}, \code{survival_event}, \code{survival_time},
and \code{case_tag} updated (samples with NA's in the case_tag vector deleted).
}
\description{
Simultaneous checking the \code{survival_time}, \code{survival_event}
and \code{case_tag} introduced in the \code{GSSTDA} object.
If there are NAs in the case_tag, the corresponding samples are removed in
\code{full_data}, \code{survival_time} and \code{survival_event}.
Function used internally in the funtions \code{gene_selection}, if the
object is not of class \code{dsga_object},  and \code{gsstda}.
}
