% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSSTDA_objects.R
\name{gene_selection.default}
\alias{gene_selection.default}
\title{gene_selection_classes.default}
\usage{
\method{gene_selection}{default}(
  data_object,
  survival_time,
  survival_event,
  gene_select_surv_type,
  percent_gen_select_for_fun_filt,
  gene_select_mapper_metric,
  percent_gen_select_for_mapper,
  na.rm = TRUE
)
}
\arguments{
\item{data_object}{Object with:
\itemize{
\item full_data Input matrix whose columns correspond to the patients and
rows to the genes.
\item case_tag Character vector of the same length as the number of
columns of \code{full_data}. Patients must be in the same order as in
\code{full_data}. It must be indicated for each patient whether its
sample is from pathological or healthy tissue. One value should be used to
indicate whether the patient's sample is healthy and another value should
be used to indicate whether the patient's sample is pathological.
The user will then be asked which one indicates whether the patient is
healthy. Only two values are valid in the vector in total.
}}

\item{survival_time}{Numerical vector of the same length as the number of
columns of \code{full_data}. In addition, the patients must be in the same
order as in \code{full_data}. For the patients whose sample is pathological
should be indicated the time between the disease diagnosis and event
(death, relapse or other). If the event has not occurred, it should be
indicated the time until the end of follow-up. Patients whose sample is
from healthy tissue must have an NA value}

\item{survival_event}{Numerical vector of the same length as the number of
columns of \code{full_data}. Patients must be in the same order as in
\code{full_data}. For the the patients with pathological sample should
be indicated whether the event has occurred (1) or not (0). Only these
values are valid and healthy patients must have an NA value.}

\item{gene_select_surv_type}{Option. Options on how to select the genes to be
used in the mapper. Select the "Abs" option, which means that the
genes with the highest absolute value are chosen, or the
"Top_Bot" option, which means that half of the selected
genes are those with the highest value (positive value, i.e.
worst survival prognosis) and the other half are those with the
lowest value (negative value, i.e. best prognosis). "Top_Bot" default option.}

\item{percent_gen_select_for_fun_filt}{Percentage (from zero to one hundred)
of genes to be selected to be used in the calculation of the values of the
filter function. 1 default option.}

\item{gene_select_mapper_metric}{Gene selection criteria for Mapper. Choose as
selection criteria between:
\itemize{
\item "mad": those genes in the disease component matrix with the highest mean
absolute deviation will be selected
\item "sd": those genes with the highest standard deviation will be selected
\item “iqr": those genes with the highest interquartile range will be selected
\item "mean_sd": to choose genes with high mean and standard deviation simultaneously
\item "sd_surv": these option selects genes for mapper based on the product of standard
deviation of the genes in the disease component matrix plus one times
the Z score obtained by fitting a cox proportional hazard model to the level
of each gene.
}}

\item{percent_gen_select_for_mapper}{Percentage (from zero to one hundred)
of genes to be selected to be used in Mapper. 5 default option.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows. TRUE default
option.}
}
\value{
A \code{gene_selection_object}. It contains:
\itemize{
\item the matrix with which the gene selection has been performed without NAN's
values (\code{data}). It is the \code{matrix_disease_component} in case it has been
performed from a \code{dsga_object} or \code{full_data} in the opposite case.
\item the \code{cox_all_matrix} (a matrix with the results of the application of
proportional hazard models: with the regression coefficients, the odds ratios,
the standard errors of each coefficient, the Z values (coef/se_coef) and
the p-values for each Z value)
\item a vector with the name of the selected genes for mapper
(\code{genes_selected_mapper}) and for the calculation of the values of the
filter function (\code{genes_selected_fun_filt}).
\item \code{case_genes_disease_component}: the matrix of disease components with
only the rows of the selected genes and only the columns of the pathological
(case) samples. This matrix is one of the inputs for the \code{mapper} function.
\item and the vector of the values of the filter function of pathological samples
\code{filter_values}.
}
}
\description{
Private function to select Gene without dsga process
}
\examples{
\donttest{
data_object <- list("full_data" = full_data, "case_tag" = case_tag)
class(data_object) <- "data_object"
gene_selection_obj <- gene_selection(data_object, survival_time,
                                     survival_event,
                                     gene_select_surv_type ="top_bot",
                                     percent_gen_select_for_fun_filt=1,
                                     gene_select_mapper_metric="mad",
                                     percent_gen_select_for_mapper=5)}
}
