% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSSTDA_objects.R
\name{gene_selection}
\alias{gene_selection}
\title{Gene selection and filter function}
\usage{
gene_selection(
  data_object,
  gene_select_surv_type = "Top_Bot",
  percent_gen_select_for_fun_filt = 1,
  gene_select_mapper_metric = "mad",
  percent_gen_select_for_mapper = 5,
  na.rm = TRUE
)
}
\arguments{
\item{data_object}{Object with:
\itemize{
\item full_data Input matrix whose columns correspond to the patients and
rows to the genes.
\item survival_time Numerical vector of the same length as the number of columns
of \code{full_data}. In addition, the patients must be in the same
order as in \code{full_data}. For the patients whose sample is pathological
should be indicated the time between the disease diagnosis and event
(death, relapse or other). If the event has not occurred, it should be
indicated the time until the end of follow-up. Patients whose sample is
from healthy tissue must have an NA value
\item survival_event Numerical vector of the same length as the number of
columns of \code{full_data}. Patients must be in the same order as in
\code{full_data}. For the the patients with pathological sample should
be indicated whether the event has occurred (1) or not (0). Only these
values are valid and healthy patients must have an NA value.
\item case_tag Character vector of the same length as the number of
columns of \code{full_data}. Patients must be in the same order as in
\code{full_data}. It must be indicated for each patient whether its
sample is from pathological or healthy tissue. One value should be used to
indicate whether the patient's sample is healthy and another value should
be used to indicate whether the patient's sample is pathological.
The user will then be asked which one indicates whether the patient is
healthy. Only two values are valid in the vector in total.
}}

\item{gene_select_surv_type}{Option. Options on how to select the genes to be
used in the calculation of the values of the filter function (and in the gene
selection for Mapper if the option "sd_surv" has been chosen) Select the
"Abs" option, which means that the genes with the highest absolute value are
chosen, or the "Top_Bot" option, which means that half of the selected
genes are those with the highest value (positive value, i.e.
worst survival prognosis) and the other half are those with the
lowest value (negative value, i.e. best prognosis). "Top_Bot" default option.}

\item{percent_gen_select_for_fun_filt}{Percentage (from zero to one hundred)
of genes to be selected to be used in the calculation of the values of the
filter function. 1 default option.}

\item{gene_select_mapper_metric}{Gene selection criteria for Mapper. Choose as
selection criteria between:
\itemize{
\item "mad": those genes in the disease component matrix with the highest mean
absolute deviation will be selected
\item "sd": those genes with the highest standard deviation will be selected
\item “iqr": those genes with the highest interquartile range will be selected
\item "mean_sd": to choose genes with high mean and standard deviation simultaneously
\item "sd_surv": these option selects genes for mapper based on the product of standard
deviation of the genes in the disease component matrix plus one times
the Z score obtained by fitting a cox proportional hazard model to the level
of each gene. "mad" default option.
}}

\item{percent_gen_select_for_mapper}{Percentage (from zero to one hundred)
of genes to be selected to be used in Mapper. 5 default option.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows. TRUE default
option.}
}
\value{
A \code{gene_selection_object}. It contains:
\itemize{
\item the \code{full_data} without NAN's values (\code{data})
\item the \code{cox_all_matrix} (a matrix with the results of the application of
proportional hazard models: with the regression coefficients, the odds ratios,
the standard errors of each coefficient, the Z values (coef/se_coef) and
the p-values for each Z value)
\item a vector with the name of the selected genes
\item the matrix of disease components with only the rows of the selected genes
(\code{genes_disease_component})
\item and the vector of the values of the filter function.
}
}
\description{
Gene selection and calculation of filter function values.
After fitting a Cox proportional hazard model to each gene, this function
makes a selection of genes according to both their variability within
the database and their relationship with survival. Subsequently, with the
genes selected, the values of the filtering functions are calculated for
each patient. The filter function allows to summarise each vector of each
individual in a single data. This function takes into account the survival
associated with each gene. In particular, the implemented filter function
performs the vector magnitude in the Lp norm (as well as k powers
of this magnitude) of the vector resulting of weighting each element of
the column vector by the Z score obtained in the cox proportional
hazard model.
}
\examples{
\donttest{
data_object <- list("full_data" = full_data, "survival_time" = survival_time,
"survival_event" = survival_event, "case_tag" = case_tag)
class(data_object) <- "data_object"
gene_selection_obj <- gene_selection(data_object,
gene_select_surv_type ="top_bot", percent_gen_select_for_fun_filt=1,
gene_select_mapper_metric="mad", percent_gen_select_for_mapper=5)}
}
