% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Funct.R
\name{gene_selection_fun_filt}
\alias{gene_selection_fun_filt}
\title{Gene selection for filter function based on relationship to survival.}
\usage{
gene_selection_fun_filt(cox_all_matrix, gene_select_surv_type, num_gen_select)
}
\arguments{
\item{cox_all_matrix}{Output from the \code{cox_all_genes} function. Data.frame with
information on the relationship between genes and survival.}

\item{gene_select_surv_type}{Option. Select the "Abs" option, which means that the
genes with the highest absolute value are chosen, or the
"Top_Bot" option, which means that half of the selected
genes are those with the highest value (positive value, i.e.
worst survival prognosis) and the other half are those with the
lowest value (negative value, i.e. best prognosis).}

\item{num_gen_select}{Number of genes to be selected (those with the highest
Z score).}
}
\value{
Character vector with the names of the selected genes.
}
\description{
It selects genes for mapper filter function based on Z score obtained by
fitting a cox proportional hazard model to the level of each gene.
}
