% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_selection.R
\name{gene_selection_mapper_by_variability}
\alias{gene_selection_mapper_by_variability}
\title{Gene selection for partial clustering in Mapper based on variability.}
\usage{
gene_selection_mapper_by_variability(
  case_disease_component,
  num_gen_select,
  gene_select_mapper_metric
)
}
\arguments{
\item{case_disease_component}{Disease component matrix (output of the function
\code{generate_disease_component}) having selected only the columns
belonging to disease samples. The names of the rows must be the names of the
genes.}

\item{num_gen_select}{Number of genes to be selected (those with higher values
of the chosen variability measure).}

\item{gene_select_mapper_metric}{Measures of variability to choose as gene selection
criteria for Mapper. Choose as selection criteria between "mad" (mean
absolute deviation), "sd" (standard deviation), “iqr” (interquartile range)
and "mean_sd" (to choose genes with high mean and standard deviation
simultaneously).}
}
\value{
Character vector with the names of the selected genes.
}
\description{
It selects genes for mapper based on different measures of variability.
}
