% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Mapper.R
\name{list_to_df}
\alias{list_to_df}
\title{list_to_df}
\usage{
list_to_df(list_obj, columns_names = c("col1", "col2"))
}
\arguments{
\item{list_obj}{List of vectors that will be transformed into a data.frame.
The elements of the list must be named.}

\item{columns_names}{Optional. Names of the columns in the output data frame.}
}
\value{
Output data.frame. Each row contains each element of each vector
that was part of the list with the name that said vector received in the list.
}
\description{
Auxiliary function. It transforms a list into a two-column
data frame. The first column contains the names of the elements in the list
and the second column contains the values of each element in the list.
Each row of the resulting data.frame contains each element of each vector
that was part of the list with the name that said vector received in the
list.
}
