% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSSTDA_objects.R
\name{mapper}
\alias{mapper}
\title{Mapper object}
\usage{
mapper(
  data,
  filter_values,
  num_intervals = NULL,
  percent_overlap = 40,
  type_covering = "uniform",
  distance_type = "correlation",
  clustering_type = "hierarchical",
  num_bins_when_clustering = 8,
  linkage_type = "ward.D",
  optimal_clustering_mode = NA,
  silhouette_threshold = 0.25,
  dim_reduction = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{Input matrix whose columns correspond to the individuals
and rows to the features.}

\item{filter_values}{Vector obtained after applying the filtering function
to the input matrix, i.e, a vector with the filtering function
values for each included sample.}

\item{num_intervals}{Number of intervals used to create the first sample
partition based on filtering values. By default the root of the number of
individuals included as input to Mapper.}

\item{percent_overlap}{Percentage of overlap between intervals. Expressed
as a percentage. 40 default option.}

\item{type_covering}{Parameter to choose how to construct the covering.
Choose between "classical" or "uniform". If the ‘classic’ option is
selected, the covering is formed by splitting the range of filter function
values into overlapping intervals of the same length. If the ‘uniform’ option
is selected, the covering will be formed with overlapping intervals of the
filter function values containing approximately the same number of
individuals. "uniform" default option.}

\item{distance_type}{Type of distance to be used for clustering.
Choose between correlation ("correlation") and euclidean ("euclidean"). "correlation"
default option.}

\item{clustering_type}{Type of clustering method. Choose between
"hierarchical" and "PAM" (“partition around medoids”) options.
"hierarchical" default option.}

\item{num_bins_when_clustering}{Number of bins to generate the
histogram employed by the standard optimal number of cluster finder
method. Parameter not necessary if the "optimal_clustering_mode" option
is "silhouette" or the "clustering_type" is "PAM". 8 default option.}

\item{linkage_type}{Linkage criteria used in hierarchical clustering.
Choose between "single" for single-linkage clustering, "complete" for
complete-linkage clustering, "average" for average linkage clustering
(or UPGMA) or "ward.D" for Ward's method. Only necessary for hierarchical
clustering. "ward.D" default option.}

\item{optimal_clustering_mode}{Method for selection optimal number of
clusters. It is only necessary if the chosen type of algorithm is
hierarchical. In this case, choose between "standard" (the method used
in the original mapper article) or "silhouette". In the case of the PAM
algorithm, the method will always be "silhouette".}

\item{silhouette_threshold}{Minimum value of \eqn{\overline{s}}{s-bar} that a set of
clusters must have to be chosen as optimal. Within each interval of the
filter function, the average silhouette values \eqn{\overline{s}}{s-bar} are computed
for all possible partitions from $2$ to $n-1$, where $n$ is the number of
samples within a specific interval. The $n$ that produces the highest value
of \eqn{\overline{s}}{s-bar} and that exceeds a specific threshold is selected as the
optimum number of clusters. If no partition produces an \eqn{\overline{s}}{s-bar}
exceeding the chosen threshold, all samples are then assigned to a unique
cluster. The default value is $0.25$. The threshold of $0.25$ for
\eqn{\overline{s}}{s-bar} has been chosen based on standard practice, recognizing it
as a moderate value that reflects adequate separation and cohesion within
clusters.}

\item{dim_reduction}{Boolean. Parameter indicating whether or not centering,
scaling and dimensionality reduction of the data is wanted prior to partial
clustering in Mapper. In dimensionality reduction, principal components that explain at least 80\% of the
variance are chosen. This process is performed on each subset composed by
the columns of the individuals that are part of each interval. TRUE indicates
that centering, scaling and dimensionality reduction will be performed. FALSE
means that this step will be omitted and clustering will be performed without
transforming the data.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.
If \code{FALSE}, an error occurs in case of \code{NA} rows. TRUE default
option.}
}
\value{
A \code{mapper_obj} object. It contains the values of the intervals
(interval_data), the samples included in each interval (sample_in_level),
information about the cluster to which the individuals in each interval
belong (clustering_all_levels), a list including the individuals contained
in each detected node (node_samples), a list including the nodes contained
in each interval (nodes_in_lev), their size (node_sizes), the
average of the filter function values of the individuals of each node
(node_average_filt) and the adjacency matrix linking the nodes (adj_matrix).
}
\description{
TDA are persistent homology and mapper. Persistent homology
borrows ideas from abstract algebra to identify particular aspects
related to the shape of the data such as the number of connected
components and the presence of higher-dimensional holes, whereas
mapper condenses the information of high-dimensional datasets into
a combinatory graph or simplicial complex that is referred to as
the skeleton of the dataset. This implementation is the mapper of one
dimension, i.e. using only one filter function value.
}
\examples{
\donttest{
control_tag_cases <- which(case_tag == "NT")
gene_selection_object <- gene_selection_(full_data, survival_time, survival_event,
control_tag_cases, gene_select_surv_type ="top_bot", num_gen_select_for_fun_filt = 200,
gene_select_mapper_metric="mad", num_gen_select_for_mapper = 1000)

mapper_object <- mapper(data = gene_selection_object[["case_genes_disease_component"]],
filter_values = gene_selection_object[["filter_values"]],
num_intervals = 5,
percent_overlap = 40, type_covering = "uniform",
distance_type = "correlation", clustering_type = "hierarchical",
linkage_type = "single")}
}
