% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_D_Mapper.R
\name{nodes_in_levels}
\alias{nodes_in_levels}
\title{Nodes included in each interval of the covering.}
\usage{
nodes_in_levels(clust_all_levels_list)
}
\arguments{
\item{clust_all_levels_list}{A list with information on the levels
obtained from the \code{clust_all_levels} function.}
}
\value{
List of character vectors where each one contains the names of the
nodes resulting from the clustering corresponding to each interval of the
covering.
}
\description{
Creates a list in which each element is a vector that collects the nodes
resulting from the clustering for each interval of the covering.
}
