% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Mapper.R
\name{percent_table}
\alias{percent_table}
\title{percent_table}
\usage{
percent_table(var_1, var_2)
}
\arguments{
\item{var_1}{First variable, corresponding to the rows in the contingency
table}

\item{var_2}{Second variable, corresponding to the columns in the contingency
table}
}
\value{
Contingency table of the two variables in percentage.
}
\description{
Auxiliary function. It creates a contingency table with the
percentages of the categories of two variable of interest (columns) in each
of the nodes of the mapper graph (rows).
}
