% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Mapper.R
\name{percent_table_nodes}
\alias{percent_table_nodes}
\title{percent_table_nodes}
\usage{
percent_table_nodes(mapper_object, var_of_interest)
}
\arguments{
\item{mapper_object}{A list produced as an output of the \code{one_D_Mapper}
function.}

\item{var_of_interest}{Vector of type "character" or "factor" that collects
the variable of interest. It must have as names the names of the samples
contained in the mapper object.}
}
\value{
Contingency table in percentages showing the relative frequencies of
the variable of interest (columns) in each of the nodes of the mapper
graph (rows).
}
\description{
It creates a contingency table showing the relative frequencies of
the variable of interest in each of the nodes of the mapper
graph.
}
