% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Mapper.R
\name{plot_mapper_with_pie_chart}
\alias{plot_mapper_with_pie_chart}
\title{Plot Mapper with Pie Charts}
\usage{
plot_mapper_with_pie_chart(
  mapper_object,
  var_of_interest,
  name_var_of_interest = "Variable of interest",
  show_node_names = FALSE
)
}
\arguments{
\item{mapper_object}{A list produced as an output of the \code{one_D_Mapper}
function.}

\item{var_of_interest}{Vector of type "character" or "factor" that collects
the variable of interest. It must have as names the names of the samples
contained in the mapper object. Therefore, within a GSSTDA, we recommend that
you first select only those items corresponding to pathological samples.}

\item{name_var_of_interest}{Name of the variable of interest to be displayed
in the legend. "Variable of interest" by default.}

\item{show_node_names}{Option to indicate whether node labels should be
displayed (\code{TRUE}) or not (\code{FALSE}). By default, it takes the value
\code{FALSE}. Any value other than \code{FALSE} will be interpreted as
\code{TRUE}.}
}
\value{
Mapper plot using \code{igraph} in which each node
is represented by a pie chart showing the proportion of each category of a
variable of interest in each node.
}
\description{
It reproduces the graph of a Mapper object in which each node
is represented by a pie chart showing the proportion of each category of a
variable of interest in each node.
}
\examples{
\donttest{
# Create data object
data_object <- list("full_data" = full_data, "case_tag" = case_tag)
class(data_object) <- "data_object"

#Select gene from data object
gene_selection_object <- gene_selection(data_object, survival_time,
                                        survival_event,
                                        gene_select_surv_type = "Top_Bot",
                                        percent_gen_select_for_fun_filt = 1,
                                        gene_select_mapper_metric = "mad",
                                        percent_gen_select_for_mapper = 5)

mapper_object <- mapper(data = gene_selection_object[["case_genes_disease_component"]],
filter_values = gene_selection_object[["filter_values"]],
num_intervals = 5, percent_overlap = 40,
type_covering = "uniform", distance_type = "correlation",
clustering_type = "hierarchical",
linkage_type = "single")

# We want to represent the percentage of mortalities for each node.
names(survival_event) <- colnames(full_data)
var_of_interest <- survival_event[case_tag == "T"]
plot_mapper_with_pie_chart(mapper_object, var_of_interest, "Exitus")}
}
